train split to integer

+++ data/dataprocess.py
@@ -29,7 +29,7 @@ class DataProcess(torch.utils.data.Dataset):
 
             # Split into train set
             n_images = len(self.de_paths)
-            train_split = n_images * 0.8
+            train_split = int(n_images * 0.8)
             print('TRAIN SPLIT: ', train_split)
             self.de_paths = self.de_paths[:train_split]
             self.st_paths = self.st_paths[:train_split]
